+++ docs/CHANGE_LOG.md
@@ -1,4 +1,25 @@
 # Heat.js - Change Log:
 
+## Version 0.2.0:
+
+#### **Binding Options:**
+- Added a new binding option called "dayToolTipText", which states the tooltip text that should be shown for a day.
+
+#### **Binding Options - Custom Triggers:**
+- Added a new binding option called "onRefresh", which states an event that should be triggered when a heatmap is refreshed.
+- Added a new binding option called "onBeforeRender", which states an event that should be triggered before an element is rendered.
+- Added a new binding option called "onRenderComplete", which states an event that should be triggered after an element is rendered.
+
+#### **Public Functions:**
+- Added new public function "removeDate()", which will remove a specific date from a heat map (or decrease the date if more than one exists).
+- Added new public function "reset()", which will remove all the dates from a heat map.
+
+#### **Documentation:**
+- Improved documentation in the main README files.
+- Fixed some spelling/grammar mistakes.
+
+<br>
+
+
 ## Version 0.1.0:
 - Everything :)